ref(replays): Enhance user badge with live indicator and improve layout
- Integrated a live indicator in the ReplayDetailsUserBadge component to show when a replay is active.
- Updated layout by using the Flex component for better alignment and presentation of user details.
- Removed outdated comments and streamlined the LiveIndicator styling for improved clarity and performance.
